draw分割
