+++ src/main.c
@@ -6,52 +6,11 @@
 /*   By: ttsubo <ttsubo@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/22 12:48:48 by ttsubo            #+#    #+#             */
-/*   Updated: 2025/03/20 20:25:50 by ttsubo           ###   ########.fr       */
+/*   Updated: 2025/03/20 20:36:51 by ttsubo           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "so_long.h"
-
-// TODO: TOO_MANY_LINES
-int	draw(void *param)
-{
-	t_game_data	*data;
-	t_glx		*glx;
-	t_pos		draw_pos;
-	int			i;
-
-	glx = get_glx();
-	data = (t_game_data *)param;
-	draw_pos = (t_pos){.x = 0, .y = 32};
-	glx->cls();
-	i = 0;
-	while (data->map->data[i])
-	{
-		if (data->map->data[i] == '0')
-			glx->put_img(0, draw_pos);
-		if (data->map->data[i] == '1')
-			glx->put_img(1, draw_pos);
-		if (data->map->data[i] == 'C')
-			glx->put_img(2, draw_pos);
-		if (data->map->data[i] == 'E')
-			glx->put_img(3, draw_pos);
-		if (data->map->data[i] == 'P')
-			glx->put_img(4, draw_pos);
-		if (data->map->data[i] == '\n')
-		{
-			draw_pos.y += 16;
-			draw_pos.x = 0;
-			i++;
-			continue ;
-		}
-		i++;
-		draw_pos.x += 16;
-	}
-	glx->text("move:", (t_pos){16,16}, GLX_COLOR_CORN_FLOWER_BLUE);
-	glx->text(ft_itoa(data->count), (t_pos){48,16}, GLX_COLOR_CORN_FLOWER_BLUE);
-	glx->text("so_long", (t_pos){250, 16}, GLX_COLOR_MINT_GREEN);
-	return (0);
-}
 
 int	clean(void *param)
 {
